+++ blogs/ipo.mdx
@@ -1,6 +1,7 @@
 ---
 title: Easter IPO 🍾💃💵🤖🚀
 image: shop-banner.png
+ipo: true
 ---
 
 <Title style={{ textAlign: "center" }}>
