+++ containers/IG/Posts.tsx
@@ -54,16 +54,26 @@ type PostsProps = {
 		numberOfView?: string;
 		numberOflike?: string;
 		numberOfcomment: string;
-	};
+		onClick?: () => void;
+	}[];
 	avatar: string;
 	username: string;
 };
 
-const Posts = ({ data, avatar, username }: PostsProps) => {
+const BlogSlider = ({ data, avatar, username }: PostsProps) => {
 	const [postLimit, setPostLimit] = useState(9);
 	const [currentPost, setCurrentPost] = useState(1);
 	const [visible, setVisible] = useState(false);
+	const [loading, setLoading] = useState(false);
 	const [direction] = useDirection();
+
+	const handleLoadMore = () => {
+		setLoading(true);
+		setTimeout(() => {
+			setLoading(false);
+			setPostLimit(postLimit + 3);
+		}, 600);
+	};
 
 	const handleModal = (id: number) => {
 		setCurrentPost(id);
@@ -82,15 +92,18 @@ const Posts = ({ data, avatar, username }: PostsProps) => {
 		return { __html: data };
 	};
 
-	let newData: any = data;
+	let newData: any;
+	data.forEach((item) => {
+		if (parseInt(item.id) === currentPost) {
+			newData = item;
+		}
+	});
 
 	return (
 		<Wrapper>
 			<Container>
-
-				{/* Post Image here */}
 				<Row>
-					{/* {data.slice(0, postLimit).map((post) => (
+					{data.slice(0, postLimit).map((post) => (
 						<Col sm={6} md={4} key={`post-key${post.id}`}>
 							<InstagramCard
 								style={{ marginBottom: '20px' }}
@@ -102,8 +115,38 @@ const Posts = ({ data, avatar, username }: PostsProps) => {
 								onClick={() => handleModal(parseInt(post.id))}
 							/>
 						</Col>
-					))} */}
+					))}
 				</Row>
+
+				<Block
+					paddingTop={['10px', '10px', '20px']}
+					display="flex"
+					alignItems="center"
+					justifyContent="center"
+				>
+					{data.length > postLimit ? (
+						<Button
+							isLoading={loading}
+							onClick={handleLoadMore}
+							kind="secondary"
+							size="large"
+							overrides={{
+								BaseButton: {
+									style: ({ $theme }) => {
+										return {
+											height: '48px',
+											...$theme.typography.font250,
+										};
+									},
+								},
+							}}
+						>
+							Check {data.length - postLimit} more posts
+						</Button>
+					) : (
+						'No more post'
+					)}
+				</Block>
 
 				<Modal
 					onClose={() => {
@@ -142,16 +185,17 @@ const Posts = ({ data, avatar, username }: PostsProps) => {
 						},
 					}}
 				>
+					{currentPost > 1 && (
+						<PrevButton onClick={handlePrevPost}>
+							<IoIosArrowBack />
+						</PrevButton>
+					)}
 
-					<PrevButton>
-						<IoIosArrowBack />
-					</PrevButton>
-
-
-					<NextButton onClick={handleNextPost}>
-						<IoIosArrowForward />
-					</NextButton>
-
+					{currentPost < data.length && (
+						<NextButton onClick={handleNextPost}>
+							<IoIosArrowForward />
+						</NextButton>
+					)}
 
 					<ContentWrapper>
 						<Media>
@@ -199,26 +243,26 @@ const Posts = ({ data, avatar, username }: PostsProps) => {
 								<CommentWrapper>
 									{newData.comments !== undefined && newData.comments.length > 0
 										? newData.comments.map((item: any) => (
-											<Comment
-												key={`comment-key${item.id}`}
-												role={item.role}
-												avatar={item.avatar}
-												name={item.username}
-												content={item.comment}
-												handleLike={() =>
-													console.log(
-														'Write like function for post.',
-														newData.id
-													)
-												}
-												handleReply={() =>
-													console.log(
-														'Write reply function for post.',
-														newData.id
-													)
-												}
-											/>
-										))
+												<Comment
+													key={`comment-key${item.id}`}
+													role={item.role}
+													avatar={item.avatar}
+													name={item.username}
+													content={item.comment}
+													handleLike={() =>
+														console.log(
+															'Write like function for post.',
+															newData.id
+														)
+													}
+													handleReply={() =>
+														console.log(
+															'Write reply function for post.',
+															newData.id
+														)
+													}
+												/>
+										  ))
 										: ''}
 								</CommentWrapper>
 							</Body>
@@ -251,4 +295,4 @@ const Posts = ({ data, avatar, username }: PostsProps) => {
 	);
 };
 
-export default Posts;
+export default BlogSlider;
